requirements.txt
notebook
Pillow
ipywidgets
setuptools-git